Move doc comment parsing into solang-parser
This is so it can be shared with forge-doc.
