+++ solang-parser/src/lib.rs
@@ -4,6 +4,7 @@ use diagnostics::Diagnostic;
 use lalrpop_util::ParseError;
 
 pub mod diagnostics;
+pub mod doccomment;
 mod lexer;
 pub mod pt;
 #[cfg(test)]
